Separate tests to different files
